requirements.txt
# ECG Monitoring Dashboard Requirements
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.13.0
scipy>=1.10.0
joblib>=1.2.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
pyserial>=3.5
